`sappho-east`: split out effects types into individual mods.
